Renamed dirs again and updated relavent files :p MainApplication to GreenEnergySearch
