+++ src/Clean-Energy-PA-Site/GreenEnergySearch/templates/GreenEnergySearch/home.html
@@ -1,0 +1,41 @@
+{% extends 'GreenEnergySearch/base.html' %}
+<div class="center">
+  <h4>Did you know that PA state law allows you to switch energy suppliers from PECO?<br />And - you can actually switch to 100% renewable</h4>
+
+  {% block content %}
+    <div class="container">
+      <div class="row justify-content-center">
+        <div class="col-md-8">
+          <div class="text-center mb-4">
+            <h4 class="text-primary">Did you know that PA state law allows you to switch energy suppliers from PECO?</h4>
+            <h4 class="text-primary">And you can actually switch to 100% renewable.</h4>
+          </div>
+
+          <form class="text-center">
+            <div class="row mb-3">
+              <div class="col-md-6 offset-md-3">
+                <div class="form-group">
+                  <input type="zipcode" class="form-control" id="zipcode" placeholder="Enter Zip Code" />
+                </div>
+              </div>
+            </div>
+            <div class="row mb-3">
+              <div class="col-md-6 offset-md-3">
+                <div class="text-center">
+                  <button type="submit" class="btn btn-success btn-block">Search Cheaper Green Energy</button>
+                </div>
+              </div>
+            </div>
+          </form>
+
+          <div class="text-center mt-4">
+            <h4 class="text-muted">Clean Energy PA is an open-source application made by graduate students at Penn State.</h4>
+
+            <p class="lead">The service is free and aims to make it easier to sign up and track suppliers, find an initial green rate below PECO's price, keep track of your contract lengths, and alert you when cheaper options are available or when your contract is about to expire.</p>
+            <p class="lead">It's 100% free, the code is open-source, and all data is kept private.</p>
+          </div>
+        </div>
+      </div>
+    </div>
+  {% endblock %}
+</div>
