+++ src/Clean-Energy-PA-Site/UserRegistration/templates/UserRegistration/register.html
@@ -1,4 +1,4 @@
-{% extends 'MainApplication/base.html' %}
+{% extends 'GreenEnergySearch/base.html' %}
 
 {% block title %}
   Create an Account
