+++ src/Clean-Energy-PA-Site/Website/urls.py
@@ -21,6 +21,6 @@ from UserRegistration import views as v
 urlpatterns = [
     path("admin/", admin.site.urls),
     path("register/", v.register, name="register"),
-    path("", include("MainApplication.urls")),
+    path("", include("GreenEnergySearch.urls")),
     path("", include("django.contrib.auth.urls")),
 ]
